+++ _playground/bundle.js
@@ -7,17 +7,24 @@ const POST_CSS_PLUGINS = [precss()];
 
 const fuseBox = FuseBox.init({
     homeDir: "_playground/ts",
-    sourceMap: {
-        bundleReference: "./sourcemaps.js.map",
-        outFile: "_playground/_build/sourcemaps.js.map",
-    },
+    // sourceMap: {
+    //     bundleReference: "./sourcemaps.js.map",
+    //     outFile: "_playground/_build/sourcemaps.js.map",
+    // },
     //globals: { jQuery: "$" },
-    cache: false,
+    cache: true,
     //globals: { default: "myLib", "wires-reactive": "Reactive" },
     outFile: "_playground/_build/out.js",
     //package: "myLib",
     //globals: { myLib: "myLib" },
     modulesFolder: "_playground/npm",
+    shim: {
+        jquery: {
+            source: "node_modules/jquery/dist/jquery.js",
+            exports: "$"
+        }
+    },
+    log: false,
     plugins: [
         build.TypeScriptHelpers(),
         build.JSONPlugin(),
@@ -28,16 +35,24 @@ const fuseBox = FuseBox.init({
         // process them all ;-)
         [build.LESSPlugin(), build.CSSPlugin()],
 
-        [build.SassPlugin(), build.CSSPlugin({ write: true })],
+        [build.SassPlugin(), build.CSSPlugin()],
 
         // All other CSS files
         [build.PostCSS(POST_CSS_PLUGINS), build.CSSPlugin()],
 
         // Add a banner to bundle output
-        build.BannerPlugin('// Hey this is my banner! Copyright 2016!')
+        build.BannerPlugin('// Hey this is my banner! Copyright 2016!'),
 
         //build.UglifyJSPlugin()
+
+        build.HTMLPlugin(),
     ]
 });
 
-fuseBox.bundle(">index.ts");
+const devServer = fuseBox.devServer(">index.ts", {
+    port: 8083,
+    emitter: (self, fileInfo) => {
+        self.socketServer.send("source-changed", fileInfo);
+    }
+});
+console.log(devServer.httpServer.app);