+++ _playground/shims/foobar.js
@@ -1,0 +1,1 @@
+window.foobar = "foo bar from library";