+++ _playground/ts/foo.ts
@@ -1,0 +1,10 @@
+export class MyFoo {
+    constructor() {
+        //console.log("HAHpppo!!!k     ...", new Date());
+    }
+}
+
+
+
+
+