+++ assets/libs/fusebox-hot-reload/index.js
@@ -1,0 +1,28 @@
+const Client = require("fusebox-websocket").SocketClient;
+
+module.exports = {
+    connect: () => {
+        if (FuseBox.isServer) {
+            return;
+        }
+        let client = new Client();
+        client.connect();
+        console.log("connecting...");
+        client.on("source-changed", (data) => {
+            console.log(`Updating "${data.path}" ...`);
+            if (data.type === "js") {
+                FuseBox.flush();
+                FuseBox.dynamic(data.path, data.content);
+                if (FuseBox.mainFile) {
+                    FuseBox.import(FuseBox.mainFile)
+                }
+            }
+            if (data.type === "css" && __fsbx_css) {
+                __fsbx_css(data.path, data.content)
+            }
+        })
+        client.on("error", (erro) => {
+            console.log(error);
+        });
+    }
+}